Extract a separate module for the emitting.
